54 frontend that loads placeholder loginregistration (#72)
* Initial react paths and component files

Installed react-router-dom and the latest Vite. Created the compnents folder and the main front end file that have inital basic returns.

* passed setToken into Account.jsx

* added curly brackets to props per requested changes

* Corrected spelling of recipe

* added headers and paragraph brackets per request

* Initial code for POST login and register

created inital functions for the login and register forms on the login page. need to clarify endpoint location and if we have certain conditional messages.

* password match, login/register don't share state

We now have password matching that disables or enables the submit button depending on user entry. I altered the states for login and register for passwords and usernames so they no longer share a state.

* Removed comments