Main page
Added main page functionality to display a random 'featured' recipe and
show all current tags as links. It works with dummy data but is not a
real get because of my issues with the call. The real data will be pass
ing through [recipes, setRecipes], then I have it set to go through the
function at line 140 which selects the featRecipe, which is what renders
on screen.