Add tabs component (#122)
* made tags component

* Main page

Added main page functionality to display a random 'featured' recipe and
show all current tags as links. It works with dummy data but is not a
real get because of my issues with the call. The real data will be pass
ing through [recipes, setRecipes], then I have it set to go through the
function at line 140 which selects the featRecipe, which is what renders
on screen.

* Fixed get call

Featured Recipe & tags are now functional

* add-tabs-component

---------

Co-authored-by: Allison Meadors <allisonmeadors@allisons-mbp.mynetworksettings.com>