79 frontend that loads placeholder other user profile page (#156)
* adding form code

* Add est time (#144)

* login styling

* added esttime

* Fixing get public user page (#142)

* added getPublicUserPage

* final commit

* 141 all user links p2 (#143)

* allrecipes page and main page have usernames on recipes now

* user link on recipe card now

* added account user links on reviews and comments

* added props to userProfile and child comps

* fixed main page links

* add-patch (#145)

* add-patch

* id fix

* code commit prior to rebase

* Conditionally rendering UserInfo based on Admin
UserProfile now has a conditional get request. If logged in as Admin it willfetch a GET request with authorization of the token and display all the userId profile data minus passwords. If not admin it will only fetch GET unauthorized even if logged in, and will only display userId's username.
The form button will only appear for Admin and preloads userData, again no passwords, but currently have start of a status switch for Admin privledges. Should effect user on next log in to recieve admin prop change. Currently this is not functioning properly.

* UserProfile PATCH form finished
UserProfile form will auto select the true false value of the user currently and PATCH to API is working.
Added checks on userInfo to only display private info is a user has token and is either admin or the current user.
Added NavButton for New Recipe to UserRecipes with the condition that the user is admin and that that is their userprofile.
corrected esttime on recipeCard.

---------

Co-authored-by: ljreaux <146141261+ljreaux@users.noreply.github.com>
Co-authored-by: jared23m <45523770+jared23m@users.noreply.github.com>